typecasting in player

+++ Western Gold/Western Gold/SFML/CodePC/MainProgram/Player.cpp
@@ -23,11 +23,11 @@ void Player::move(DeltaTime time)
 	}
 	//If player as two keys down move half speed
 	if (movementKeys > 0) {
-		int nowSpeed = speed;
+		float nowSpeed = speed;
 		if (movementKeys >= 2) {
 			nowSpeed /= 2;
 		}
-		this->moveSprite(time.dt() * speed * xDir, time.dt() * speed * yDir);
+		this->moveSprite((float)time.dt() * speed * xDir, (float)time.dt() * speed * yDir);
 	}
 }
